+++ src/raster.cpp
@@ -9,7 +9,7 @@ void rasterize_triangle(const Program& program, const UniformAttributes& uniform
 			 x2, y2, z2, 1
 		*/
 		
-		Eigen::Matrix<float,3,4> p;
+		Eigen::Matrix<double,3,4> p;
 		p.row(0) = v1.position.array()/v1.position[3];
 		p.row(1) = v2.position.array()/v2.position[3];
 		p.row(2) = v3.position.array()/v3.position[3];
@@ -35,13 +35,13 @@ void rasterize_triangle(const Program& program, const UniformAttributes& uniform
 		uy = std::min(std::max(uy,int(0)),int(frameBuffer.cols()-1));
 
 		// Build the implicit triangle representation
-		Eigen::Matrix3f A;
+		Eigen::Matrix3d A;
 		A.col(0) = p.row(0).segment(0,3);
 		A.col(1) = p.row(1).segment(0,3);
 		A.col(2) = p.row(2).segment(0,3);
 		A.row(2) << 1.0, 1.0, 1.0;
 
-		Eigen::Matrix3f Ai = A.inverse();
+		Eigen::Matrix3d Ai = A.inverse();
 
 		// Rasterize the triangle
 		for (unsigned i=lx; i<=ux; i++)
@@ -49,8 +49,8 @@ void rasterize_triangle(const Program& program, const UniformAttributes& uniform
 			for (unsigned j=ly; j<=uy; j++)
 			{
 				// The pixel center is offset by 0.5, 0.5
-				Eigen::Vector3f pixel(i+0.5,j+0.5,1);
-				Eigen::Vector3f b = Ai*pixel;
+				Eigen::Vector3d pixel(i+0.5,j+0.5,1);
+				Eigen::Vector3d b = Ai*pixel;
 				if (b.minCoeff() >= 0)
 				{
 					VertexAttributes va = VertexAttributes::interpolate(v1,v2,v3,b[0],b[1],b[2]);
@@ -86,7 +86,7 @@ void rasterize_triangles(const Program& program, const UniformAttributes& unifor
 void rasterize_line(const Program& program, const UniformAttributes& uniform, const VertexAttributes& v1, const VertexAttributes& v2, float line_thickness, FrameBuffer& frameBuffer)
 {
 		// Collect coordinates into a matrix and convert to canonical representation
-		Eigen::Matrix<float,2,4> p;
+		Eigen::Matrix<double,2,4> p;
 		p.row(0) = v1.position.array()/v1.position[3];
 		p.row(1) = v2.position.array()/v2.position[3];
